service_etl/requirements.txt
fastapi
uvicorn
requests
numpy
qdrant-client
beautifulsoup4
lxml
openai
python-dotenv